Problem: capnp compilation process must be fbp agnostic
Solution: the product of capnp encode should be the imsg
input. capnp tool shouldn't be called from rust code because
that means the fbp graph implementation is dependent on capnp at
run time.

By making the fbp graph implementation dependent on the product
of the capnp encode output the graph becomes portable. i.e.
subgraphs can be parsed in javascript and the implemantation
doesn't have to depend on capnp tool at runtime... in the
browser this is problematic.